[Refact] #12 - BottomSheet 로그인 수정
- 기존 로그인하면 SignUpViewModel에서 진행했었음. 이를 InitialViewModel로 이전
