+++ app/src/main/res/layout/fragment_login_bottomsheet.xml
@@ -3,6 +3,13 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <data>
+
+        <variable
+            name="viewModel"
+            type="org.shop.yogizogi_android.ui.view.auth.initial.InitialViewModel" />
+    </data>
+
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -60,7 +67,8 @@
                 android:id="@+id/et_login_phonenumber"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:maxLines="1" />
+                android:maxLines="1"
+                app:text="@{viewModel.phoneNumber}" />
         </com.google.android.material.textfield.TextInputLayout>
 
         <androidx.appcompat.widget.AppCompatTextView
@@ -83,6 +91,7 @@
             android:layout_marginEnd="43dp"
             android:hint="@string/login_password_hint"
             app:endIconMode="password_toggle"
+            app:errorEnabled="false"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="@id/tv_login_phonenumber"
             app:layout_constraintTop_toBottomOf="@id/tv_login_password"
@@ -94,7 +103,8 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:inputType="textPassword"
-                android:maxLines="1" />
+                android:maxLines="1"
+                app:text="@{viewModel.password}" />
         </com.google.android.material.textfield.TextInputLayout>
 
         <androidx.appcompat.widget.AppCompatTextView
